Move test data to mock folder
